Adds Clerk authentication provider, deletes next-auth, changes folder structure to keep routes better organized
