+++ app/(main)/(routes)/payments/page.tsx
@@ -1,41 +1,38 @@
-"use client";
-import axios from "axios";
-import React, { useEffect, useState } from "react";
-import { ServerResponse } from "../types";
-import { StripeContainer } from "@/components";
-import { PaymentProps } from "@/lib/types";
-
+'use client';
+import axios from 'axios';
+import React, { useEffect, useState } from 'react';
+import { ServerResponse } from '@/lib/types';
+import { StripeContainer } from '@/components';
+import { PaymentProps } from '@/lib/types';
 
 // payment display/record component
 const PaymentRecord = () => {
-  const [invoice, setInvoice] = useState<string>("");
-	let paymentprops: PaymentProps;
+  const [invoice, setInvoice] = useState<string>('');
+  let paymentprops: PaymentProps;
   /**
    * define variablwe to give us access to this data betwen API calls
    * need to transition this to webhook
    */
   const fetchInvoiceData = async (): Promise<PaymentProps> => {
     const { data } = await axios.request<PaymentProps>({
-      url: "api/salesforce",
-      method: "post",
+      url: 'api/salesforce',
+      method: 'post',
     });
-    console.log("data: ", data);
+    console.log('data: ', data);
     /**
      * need to set invoice or assign to props
      */
-		paymentprops = data;
+    paymentprops = data;
     // setInvoice(data);
     return paymentprops;
   };
 
-
   // retrieves cached invoice records and puts them in a more readble object format to display as payment records
   const redisConnect = async (): Promise<ServerResponse> => {
-		
     const { data } = await axios.request<ServerResponse>({
-			url: "/api/redis",
-			method: 'post',
-    	data: paymentprops,
+      url: '/api/redis',
+      method: 'post',
+      data: paymentprops,
     });
     console.log(data);
     return data;
@@ -49,16 +46,16 @@ const PaymentRecord = () => {
   // console.log(payments);
 
   return (
-    <div className="max-w-sm w-full lg:max-w-full lg:flex">
+    <div className='w-full max-w-sm lg:flex lg:max-w-full'>
       <button
-        className="bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow"
+        className='rounded border border-gray-400 bg-white px-4 py-2 font-semibold text-gray-800 shadow hover:bg-gray-100'
         onClick={fetchInvoiceData}
       >
         bill-bot-baggins x salesforce route
       </button>
 
       <button
-        className="bg-white hover:bg-gray-100 text-gray-800 font-semibold py-2 px-4 border border-gray-400 rounded shadow"
+        className='rounded border border-gray-400 bg-white px-4 py-2 font-semibold text-gray-800 shadow hover:bg-gray-100'
         onClick={redisConnect}
       >
         bill-bot-baggins x redis
