--- app/AdminLogin/page.tsx
@@ -1,12 +1,0 @@
-import { AdminLogin } from "./login";
-import Navbar from "@/components/Navbar";
-
-const page = () => {
-  return (
-    <>
-    <AdminLogin />
-    </>
-  )
-}
-
-export default page
