--- app/clientLogin/page.tsx
@@ -1,12 +1,0 @@
-import { ClientLogin } from "./login";
-import Navbar from "@/components/Navbar";
-
-const page = () => {
-  return (
-    <>
-    <ClientLogin />
-    </>
-  )
-}
-
-export default page
