--- app/clientRegister/page.tsx
@@ -1,12 +1,0 @@
-import { ClientRegister } from "./form";
-import Navbar from "@/components/Navbar"
-
-const page = () => {
-  return (
-    <>
-    <ClientRegister />
-    </>
-  )
-}
-
-export default page
